pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.4",
    "openai>=2.7.2",
    "pdfplumber>=0.11.8",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "werkzeug>=3.1.3",
]
